feat(item): generate social images automatically (#559)
Some websites seem to not quite like the delay caused by us checking R2 for the image, and generating one if it doesn't exist. For example, Discord refuses to show images for any Kanojo links.

A better solution for this is to pre-generate all social images, then directly returning the link to our R2 bucket as the social image for an item.

This adds a SocialMediaPreview media collection type, and defines this collection to be single file on both Movie and Person. A Job then runs on creation and update for both models. Since the collections are single-file, the preview will be replaced automatically.

Closes #559